Added board!
There's a basic board now! This took longer than it should've and than what I thought! I think I'm tired! (it took like 4 tries ugh)
Added getters to Square class
Filled the .conf file with the given board but it will be changed because the code fills the board bottom to top, but the squares are in increasing order in the file (which alternates direction each row in the "real" board)
Renamed classes to respect what the teacher wants
Renamed a bunch of final variables so they're all caps and SCREAMING AT YOU